directly pass feature queue to task_manager
The features passes to JCukeForker::Runner should implement shift.  Shift
should be a non-blocking operation that returns a feature or nil.

This will allow us to read features from a distributed queue.  Hence we could
distribute jcukeforker instances on mulitple nodes each reading from the queue
for the job.  Setting up and running jcukeforker on the remote nodes is left to the user.